add SuperDirt-like `legato`
I've made changes to Tidal-MIDI to support a cleaner `legato` much like
SuperDirt.

1) As before, duration of notes defaults to the Tidal default: 0.05
seconds, but can be overridden in the source code on a per-synth basis
2) As before, you can further override this live with the `dur`
parameter
3) As before, you can use `unit "cycle"` to get notes with a duration
according to their "slot" in the pattern, like `legato 1` in SuperDirt
4) NEW: `unit "c"` also works as an abbreviation
5) NEW: you can use the `legato` parameter for MIDI just like you would
with SuperDirt. This automatically implies `unit "c"`, you don't have
to specify it
6) NEW: Using `legato` in a pattern overrides any `dur` also present
7) NEW: The Ambika setup was using `legato` as a MIDI CC param, I've
changed the name of this to `mlegato` to avoid conflicts.